refactor: Pull out `map` of todos into a TodosContainer component
